requirements.txt
jupyter
matplotlib
pillow
requests
numpy
pandas
scikit-learn
scikit-image
torch
torchvision